Update README
Marshall, when looking through your app I see that off the bat your photo does not work. As I looke through the code and try and see what happened I see that in your code. in your master file there is no photo that shows up, yet there is code for it in the file. What needs to happen would to have a file in your resources that says photo or images and make sure that within those files it looks like photo: /images/photoname.jpg. When you do this it is a way for titanium to know that you are linking the photo in the new images folder to the app and that's where it will come from. 

Other than this the app looks good and works all the way. I was able to make sure that I could navigate all the way through back and too each page. I liked the lay out and hope that some day that you can be a GM of a team. That is where I would want to be as well (maybe we could work together!) haha. 

I liked the differences in colors and the way it all looked. good job and the thought process behind the color scheme and good luck with the rest of the class.

Also, When I tried to import your proassign4 through titanium it would not let me after importing it gave me a large screen full of errors that will not work. So I am not sure if that is something that you did or a classic titanium fault. But good job on this assignment!